LPS-88405 Move to match UserFirstNameComparator
