+++ modules/apps/change-tracking/change-tracking-api/bnd.bnd
@@ -10,4 +10,4 @@ Export-Package:\
 	com.liferay.change.tracking.model,\
 	com.liferay.change.tracking.service,\
 	com.liferay.change.tracking.service.persistence,\
-	com.liferay.change.tracking.util
+	com.liferay.change.tracking.util.comparator
